requirements.txt
GitPython
markdown
pymongo
dnspython
bs4